lineage_tool/db.py
"""
Handles all database interactions for the lineage tool.

Currently, this module contains a simulation for fetching schema information
from a Greenplum database. In a production environment, the connection details
and query logic would be implemented here.
"""

import logging
from typing import Any, Dict, Set

import psycopg2  # noqa: F401
from psycopg2.extensions import connection  # noqa: F401

logger = logging.getLogger(__name__)


def fetch_schema_from_db(table_names: Set[str]) -> Dict[str, Any]:
    """
    Connects to a Greenplum database to fetch schema information for tables.

    This function simulates a database connection and returns a hardcoded schema
    for a predefined set of tables. The real implementation would query
    `information_schema.columns`.

    Args:
        table_names: A set of table names to look up in the database.

    Returns:
        A dictionary representing the database schema in a format compatible
        with sqlglot's MappingSchema.
    """
    if not table_names:
        return {}

    logger.info("--- SIMULATING DATABASE SCHEMA FETCHING ---")
    logger.info(f"--- Tables to look up: {table_names} ---")

    schema_map: Dict[str, Any] = {
        "my_gp_db": {
            "raw_data": {
                "users": {"id": "INT", "name": "VARCHAR", "email": "VARCHAR"},
                "orders": {
                    "id": "INT",
                    "user_id": "INT",
                    "amount": "DECIMAL",
                    "order_date": "DATE",
                },
            },
            "analytics": {"user_orders": {"user_id": "INT", "total_orders": "BIGINT"}},
        }
    }
    logger.info("--- Simulation complete. Using mock schema. ---")
    return schema_map