pyproject.toml
[project]
name = "data-lineage-ui"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "agate==1.7.1",
    "annotated-types==0.6.0",
    "appdirs==1.4.4",
    "attrs==23.2.0",
    "babel==2.14.0",
    "certifi==2024.2.2",
    "chardet==5.2.0",
    "charset-normalizer==3.3.2",
    "cli-helpers[styles]==2.4.0",
    "click==8.1.7",
    "colorama==0.4.6",
    "configobj==5.0.9",
    "daff==1.3.46",
    "dbt-adapters==1.14.5",
    "dbt-common==1.18",
    "dbt-core==1.9.4",
    "dbt-duckdb==1.9.1",
    "dbt-extractor==0.5.1",
    "dbt-postgres>=1.9.0",
    "dbt-semantic-interfaces==0.7.4",
    "deepdiff==7.0.1",
    "diff-cover==8.0.3",
    "duckcli==0.2.1",
    "duckdb==1.2.0",
    "idna==3.6",
    "importlib-metadata==6.11.0",
    "iniconfig==2.0.0",
    "ipykernel>=6.29.5",
    "isodate==0.6.1",
    "jinja2==3.1.3",
    "jsonschema==4.21.1",
    "jsonschema-specifications==2023.12.1",
    "leather==0.4.0",
    "markupsafe==2.1.5",
    "mashumaro[msgpack]==3.12",
    "more-itertools==10.2.0",
    "msgpack==1.0.8",
    "networkx==3.1",
    "ordered-set==4.1.0",
    "packaging==24.0",
    "parsedatetime==2.6",
    "pathspec==0.11.2",
    "pluggy==1.4.0",
    "prompt-toolkit==3.0.50",
    "protobuf==5.29.4",
    "psycopg2-binary>=2.9.10",
    "pydantic==2.6.3",
    "pydantic-core==2.16.3",
    "pygments==2.17.2",
    "pytest==8.1.1",
    "python-dateutil==2.9.0.post0",
    "python-slugify==8.0.4",
    "pytimeparse==1.1.8",
    "pytz==2024.1",
    "pyyaml==6.0.1",
    "referencing==0.33.0",
    "regex==2023.12.25",
    "requests==2.31.0",
    "rpds-py==0.18.0",
    "six==1.16.0",
    "snowplow-tracker==1.1.0",
    "sqlfluff==2.3.5",
    "sqlglot[rs]>=27.0.0",
    "sqllineage>=1.4.6",
    "sqlparse==0.5.0",
    "tabulate[widechars]==0.9.0",
    "tblib==3.0.0",
    "text-unidecode==1.3",
    "tqdm==4.66.2",
    "typing-extensions==4.10.0",
    "urllib3==1.26.18",
    "wcwidth==0.2.13",
    "zipp==3.17.0",
]

[tool.uv.workspace]
members = [
    "react-flow-tables",
]
